refactor: extract make payment transaction
